+++ code/packages/rust/adj-lang/src/_parser_grammar.rs
@@ -33,6 +33,8 @@ pub fn parser_grammar() -> ParserGrammar {
                 GrammarElement::RuleReference { name: r#"solve_decl"#.to_string() },
                 GrammarElement::RuleReference { name: r#"check_decl"#.to_string() },
                 GrammarElement::RuleReference { name: r#"optimize_decl"#.to_string() },
+                GrammarElement::RuleReference { name: r#"dictionary_decl"#.to_string() },
+                GrammarElement::RuleReference { name: r#"define_decl"#.to_string() },
             ] },
             line_number: 22,
         },
@@ -45,7 +47,7 @@ pub fn parser_grammar() -> ParserGrammar {
                 GrammarElement::RuleReference { name: r#"term"#.to_string() },
                 GrammarElement::Repetition { element: Box::new(GrammarElement::RuleReference { name: r#"annotation"#.to_string() }) },
             ] },
-            line_number: 41,
+            line_number: 43,
         },
         GrammarRule {
             name: r#"contributes_decl"#.to_string(),
@@ -58,7 +60,7 @@ pub fn parser_grammar() -> ParserGrammar {
                 GrammarElement::RuleReference { name: r#"term"#.to_string() },
                 GrammarElement::Repetition { element: Box::new(GrammarElement::RuleReference { name: r#"annotation"#.to_string() }) },
             ] },
-            line_number: 43,
+            line_number: 45,
         },
         GrammarRule {
             name: r#"evidence"#.to_string(),
@@ -66,7 +68,7 @@ pub fn parser_grammar() -> ParserGrammar {
                 GrammarElement::RuleReference { name: r#"predicate"#.to_string() },
                 GrammarElement::RuleReference { name: r#"term"#.to_string() },
             ] },
-            line_number: 53,
+            line_number: 55,
         },
         GrammarRule {
             name: r#"predicate"#.to_string(),
@@ -81,7 +83,7 @@ pub fn parser_grammar() -> ParserGrammar {
                     ] }) },
                 GrammarElement::TokenReference { name: r#"NUMBER"#.to_string() },
             ] },
-            line_number: 55,
+            line_number: 57,
         },
         GrammarRule {
             name: r#"interacts_decl"#.to_string(),
@@ -100,7 +102,7 @@ pub fn parser_grammar() -> ParserGrammar {
                 GrammarElement::RuleReference { name: r#"term"#.to_string() },
                 GrammarElement::Repetition { element: Box::new(GrammarElement::RuleReference { name: r#"annotation"#.to_string() }) },
             ] },
-            line_number: 57,
+            line_number: 59,
         },
         GrammarRule {
             name: r#"uncertain_decl"#.to_string(),
@@ -117,7 +119,7 @@ pub fn parser_grammar() -> ParserGrammar {
                 GrammarElement::RuleReference { name: r#"term"#.to_string() },
                 GrammarElement::Repetition { element: Box::new(GrammarElement::RuleReference { name: r#"annotation"#.to_string() }) },
             ] },
-            line_number: 59,
+            line_number: 61,
         },
         GrammarRule {
             name: r#"observe_decl"#.to_string(),
@@ -125,7 +127,7 @@ pub fn parser_grammar() -> ParserGrammar {
                 GrammarElement::Literal { value: r#"observe"#.to_string() },
                 GrammarElement::RuleReference { name: r#"term"#.to_string() },
             ] },
-            line_number: 61,
+            line_number: 63,
         },
         GrammarRule {
             name: r#"query_decl"#.to_string(),
@@ -133,7 +135,7 @@ pub fn parser_grammar() -> ParserGrammar {
                 GrammarElement::TokenReference { name: r#"QUESTION"#.to_string() },
                 GrammarElement::RuleReference { name: r#"term"#.to_string() },
             ] },
-            line_number: 63,
+            line_number: 65,
         },
         GrammarRule {
             name: r#"let_decl"#.to_string(),
@@ -143,7 +145,7 @@ pub fn parser_grammar() -> ParserGrammar {
                 GrammarElement::TokenReference { name: r#"EQUALS"#.to_string() },
                 GrammarElement::RuleReference { name: r#"expr"#.to_string() },
             ] },
-            line_number: 77,
+            line_number: 79,
         },
         GrammarRule {
             name: r#"expr"#.to_string(),
@@ -157,7 +159,7 @@ pub fn parser_grammar() -> ParserGrammar {
                         GrammarElement::RuleReference { name: r#"term_expr"#.to_string() },
                     ] }) },
             ] },
-            line_number: 79,
+            line_number: 81,
         },
         GrammarRule {
             name: r#"term_expr"#.to_string(),
@@ -171,7 +173,7 @@ pub fn parser_grammar() -> ParserGrammar {
                         GrammarElement::RuleReference { name: r#"factor"#.to_string() },
                     ] }) },
             ] },
-            line_number: 81,
+            line_number: 83,
         },
         GrammarRule {
             name: r#"factor"#.to_string(),
@@ -185,7 +187,7 @@ pub fn parser_grammar() -> ParserGrammar {
                     GrammarElement::TokenReference { name: r#"RPAREN"#.to_string() },
                 ] },
             ] },
-            line_number: 83,
+            line_number: 85,
         },
         GrammarRule {
             name: r#"agg"#.to_string(),
@@ -201,7 +203,7 @@ pub fn parser_grammar() -> ParserGrammar {
                 GrammarElement::TokenReference { name: r#"IDENT"#.to_string() },
                 GrammarElement::TokenReference { name: r#"RPAREN"#.to_string() },
             ] },
-            line_number: 85,
+            line_number: 87,
         },
         GrammarRule {
             name: r#"symbol_decl"#.to_string(),
@@ -211,7 +213,7 @@ pub fn parser_grammar() -> ParserGrammar {
                 GrammarElement::TokenReference { name: r#"COLON"#.to_string() },
                 GrammarElement::RuleReference { name: r#"term"#.to_string() },
             ] },
-            line_number: 95,
+            line_number: 97,
         },
         GrammarRule {
             name: r#"constrain_decl"#.to_string(),
@@ -221,7 +223,7 @@ pub fn parser_grammar() -> ParserGrammar {
                 GrammarElement::RuleReference { name: r#"relop"#.to_string() },
                 GrammarElement::RuleReference { name: r#"expr"#.to_string() },
             ] },
-            line_number: 97,
+            line_number: 99,
         },
         GrammarRule {
             name: r#"relop"#.to_string(),
@@ -234,7 +236,7 @@ pub fn parser_grammar() -> ParserGrammar {
                 GrammarElement::TokenReference { name: r#"EQUALS"#.to_string() },
                 GrammarElement::TokenReference { name: r#"NE"#.to_string() },
             ] },
-            line_number: 99,
+            line_number: 101,
         },
         GrammarRule {
             name: r#"solve_decl"#.to_string(),
@@ -249,12 +251,12 @@ pub fn parser_grammar() -> ParserGrammar {
                     ] }) },
                 GrammarElement::TokenReference { name: r#"RBRACE"#.to_string() },
             ] },
-            line_number: 101,
+            line_number: 103,
         },
         GrammarRule {
             name: r#"check_decl"#.to_string(),
             body: GrammarElement::Literal { value: r#"check"#.to_string() },
-            line_number: 103,
+            line_number: 105,
         },
         GrammarRule {
             name: r#"optimize_decl"#.to_string(),
@@ -265,7 +267,61 @@ pub fn parser_grammar() -> ParserGrammar {
                     ] }) },
                 GrammarElement::RuleReference { name: r#"expr"#.to_string() },
             ] },
-            line_number: 110,
+            line_number: 112,
+        },
+        GrammarRule {
+            name: r#"dictionary_decl"#.to_string(),
+            body: GrammarElement::Sequence { elements: vec![
+                GrammarElement::Literal { value: r#"dictionary"#.to_string() },
+                GrammarElement::TokenReference { name: r#"IDENT"#.to_string() },
+                GrammarElement::TokenReference { name: r#"LBRACE"#.to_string() },
+                GrammarElement::Repetition { element: Box::new(GrammarElement::RuleReference { name: r#"define_decl"#.to_string() }) },
+                GrammarElement::TokenReference { name: r#"RBRACE"#.to_string() },
+            ] },
+            line_number: 124,
+        },
+        GrammarRule {
+            name: r#"define_decl"#.to_string(),
+            body: GrammarElement::Sequence { elements: vec![
+                GrammarElement::Literal { value: r#"define"#.to_string() },
+                GrammarElement::TokenReference { name: r#"IDENT"#.to_string() },
+                GrammarElement::TokenReference { name: r#"COLON"#.to_string() },
+                GrammarElement::RuleReference { name: r#"define_kind"#.to_string() },
+                GrammarElement::Repetition { element: Box::new(GrammarElement::RuleReference { name: r#"surface_clause"#.to_string() }) },
+            ] },
+            line_number: 126,
+        },
+        GrammarRule {
+            name: r#"define_kind"#.to_string(),
+            body: GrammarElement::Alternation { choices: vec![
+                GrammarElement::Literal { value: r#"hypothesis"#.to_string() },
+                GrammarElement::Sequence { elements: vec![
+                    GrammarElement::Literal { value: r#"finding"#.to_string() },
+                    GrammarElement::Optional { element: Box::new(GrammarElement::Sequence { elements: vec![
+                            GrammarElement::Literal { value: r#"values"#.to_string() },
+                            GrammarElement::TokenReference { name: r#"LBRACK"#.to_string() },
+                            GrammarElement::TokenReference { name: r#"IDENT"#.to_string() },
+                            GrammarElement::Repetition { element: Box::new(GrammarElement::Sequence { elements: vec![
+                                    GrammarElement::TokenReference { name: r#"COMMA"#.to_string() },
+                                    GrammarElement::TokenReference { name: r#"IDENT"#.to_string() },
+                                ] }) },
+                            GrammarElement::TokenReference { name: r#"RBRACK"#.to_string() },
+                        ] }) },
+                ] },
+            ] },
+            line_number: 128,
+        },
+        GrammarRule {
+            name: r#"surface_clause"#.to_string(),
+            body: GrammarElement::Sequence { elements: vec![
+                GrammarElement::Literal { value: r#"surface"#.to_string() },
+                GrammarElement::TokenReference { name: r#"STRING"#.to_string() },
+                GrammarElement::Repetition { element: Box::new(GrammarElement::Sequence { elements: vec![
+                        GrammarElement::TokenReference { name: r#"COMMA"#.to_string() },
+                        GrammarElement::TokenReference { name: r#"STRING"#.to_string() },
+                    ] }) },
+            ] },
+            line_number: 132,
         },
         GrammarRule {
             name: r#"annotation"#.to_string(),
@@ -274,7 +330,7 @@ pub fn parser_grammar() -> ParserGrammar {
                 GrammarElement::RuleReference { name: r#"locator_annotation"#.to_string() },
                 GrammarElement::RuleReference { name: r#"trust_annotation"#.to_string() },
             ] },
-            line_number: 118,
+            line_number: 140,
         },
         GrammarRule {
             name: r#"source_annotation"#.to_string(),
@@ -282,7 +338,7 @@ pub fn parser_grammar() -> ParserGrammar {
                 GrammarElement::Literal { value: r#"source"#.to_string() },
                 GrammarElement::TokenReference { name: r#"STRING"#.to_string() },
             ] },
-            line_number: 123,
+            line_number: 145,
         },
         GrammarRule {
             name: r#"locator_annotation"#.to_string(),
@@ -290,7 +346,7 @@ pub fn parser_grammar() -> ParserGrammar {
                 GrammarElement::Literal { value: r#"locator"#.to_string() },
                 GrammarElement::TokenReference { name: r#"STRING"#.to_string() },
             ] },
-            line_number: 124,
+            line_number: 146,
         },
         GrammarRule {
             name: r#"trust_annotation"#.to_string(),
@@ -298,7 +354,7 @@ pub fn parser_grammar() -> ParserGrammar {
                 GrammarElement::Literal { value: r#"trust"#.to_string() },
                 GrammarElement::RuleReference { name: r#"trust_tier"#.to_string() },
             ] },
-            line_number: 125,
+            line_number: 147,
         },
         GrammarRule {
             name: r#"trust_tier"#.to_string(),
@@ -309,7 +365,7 @@ pub fn parser_grammar() -> ParserGrammar {
                 GrammarElement::Literal { value: r#"inferred"#.to_string() },
                 GrammarElement::Literal { value: r#"unattributed"#.to_string() },
             ] },
-            line_number: 127,
+            line_number: 149,
         },
         GrammarRule {
             name: r#"term"#.to_string(),
@@ -331,7 +387,7 @@ pub fn parser_grammar() -> ParserGrammar {
                         GrammarElement::TokenReference { name: r#"RPAREN"#.to_string() },
                     ] }) },
             ] },
-            line_number: 144,
+            line_number: 166,
         },
     ],
         version: 1,